fix(phoneme_encoder): cast mask to float before residual multiply
  Multiplying by a bool mask during bfloat16 training silently upcasts x
  to float32. Use .to(x.dtype) to stay in the correct precision, consistent
  with speech_prompt_encoder and TransformerEncoderLayer.

+++ naturalspeech2/modules/phoneme_encoder.py
@@ -50,7 +50,7 @@ class PhonemeEncoder(nn.Module):
         
         phoneme_tokens_emb = self.final_norm(phoneme_tokens_emb)
 
-        phoneme_tokens_emb = phoneme_tokens_emb * phoneme_tokens_mask
+        phoneme_tokens_emb = phoneme_tokens_emb * phoneme_tokens_mask.to(phoneme_tokens_emb.dtype)
         
         return phoneme_tokens_emb # [B, P, hidden_dim]
